agregando productos al carrito

+++ back-top-shop/src/main/java/com/back_top_shop/model/TShirt.java
@@ -7,11 +7,12 @@ import lombok.*;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class TShirt {
+public class TShirt 
+{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String title;
